moved ZoomToLinkedElement method to LinkElement
